Refactoring: Change coin type representation (closes #126) (#132)
* activity: refactor dml monetary types from long to double

* bank: refactor dml monetary types from long to double

* broker: refactor dml monetary types from long to double

* car: refactor dml monetary types from long to double

* hotel: refactor dml monetary types from long to double

* tax: refactor dml monetary types from long to double

* Convert doubles to longs in module hotel

* Convert doubles to longs in module tax

* Convert doubles to longs in module car

* Fix car price (although it did not inttterfere with the test)

* FIX: activity: refactor dml monetary types from long to double

* FIX: broker: refactor dml monetary types from long to double

* FIX: bank: refactor dml monetary types from long to double

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Account.java
@@ -36,7 +36,7 @@ public class Account extends Account_Base {
 
 	}
 
-	public Operation deposit(double amount) {
+	public Operation deposit(long amount) {
 		if (amount <= 0) {
 			throw new BankException();
 		}
@@ -46,7 +46,7 @@ public class Account extends Account_Base {
 		return new OperationDeposit(this, amount);
 	}
 
-	public Operation withdraw(double amount) {
+	public Operation withdraw(long amount) {
 		if (amount <= 0 || amount > getBalance()) {
 			throw new BankException();
 		}
@@ -56,7 +56,7 @@ public class Account extends Account_Base {
 		return new OperationWithdraw(this, amount);
 	}
 
-	public Operation transfer(double amount, Account account) {
+	public Operation transfer(long amount, Account account) {
 		if (amount <= 0 || amount > getBalance()) {
 			throw new BankException();
 		}
