+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/OperationTransfer.java
@@ -6,7 +6,7 @@ import pt.ulisboa.tecnico.softeng.bank.exception.BankException;
 
 public class OperationTransfer extends OperationTransfer_Base {
     
-    public OperationTransfer(Account account, double value, String targetIBAN) {
+    public OperationTransfer(Account account, long value, String targetIBAN) {
         checkArguments(account, value, targetIBAN);
 
         setReference(account.getBank().getCode() + Integer.toString(account.getBank().getCounter()));
@@ -33,7 +33,7 @@ public class OperationTransfer extends OperationTransfer_Base {
         throw new BankException();
     }
 
-    private void checkArguments(Account account, double value, String targetIBAN){
+    private void checkArguments(Account account, long value, String targetIBAN){
         if (account == null || value <= 0 || targetIBAN == null || targetIBAN.isEmpty()) {
             throw new BankException();
         }
